initial refactor migration from page to layout
